+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Migrates sorting logic into inventoryUtils and refactor expiration date calculation and enhance sorting logic to better handle empty values 
 - Refactor sorting logic in MainContainer to utilize new sortItems utility function
 - Enhance QuickStatsBar component with default props for better usability
+- Refactor active filter count logic into a utility function for better code reuse
 
 ### Fixed